backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paperloom"
version = "0.1.0"
description = "Local-first document toolkit. Streaming OCR (GLM-OCR via Ollama), PII anonymizer, and 19 chainable PDF/Markdown/HTML tools. MCP server included."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Lucio Palmieri" }]
requires-python = ">=3.11,<3.14"
keywords = [
    "ocr", "pdf", "markdown", "anonymization", "privacy",
    "local-first", "mcp", "llm", "ollama", "agent",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup",
    "Topic :: Multimedia :: Graphics :: Capture :: Scanners",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "pypdfium2>=4.30.0",
    "pypdf>=5.1.0",
    "Pillow>=11.0.0",
    "markdown-it-py>=3.0.0",
    "reportlab>=4.2.0",
    "mcp>=1.12.0",
]

[project.optional-dependencies]
# WeasyPrint needs native libs (pango/cairo). Opt-in to avoid surprising
# `uvx paperloom` users who only need OCR.
pdf = ["weasyprint>=68.1"]
# OPF anonymizer — torch + transformers + ~4 GB checkpoint. Opt-in.
anonymizer = ["opf"]
# All extras (web UI / dev convenience).
all = ["weasyprint>=68.1", "opf"]

[project.scripts]
paperloom = "paperloom.cli:main"
paperloom-mcp = "paperloom.mcp_server:main"

[project.urls]
Homepage = "https://github.com/luciopalmieri/paperloom"
Repository = "https://github.com/luciopalmieri/paperloom"
Issues = "https://github.com/luciopalmieri/paperloom/issues"
Documentation = "https://github.com/luciopalmieri/paperloom/tree/main/doc"

[tool.uv.sources]
opf = { git = "https://github.com/openai/privacy-filter.git", rev = "main" }

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["paperloom"]

[tool.hatch.build.targets.sdist]
include = ["paperloom", "tests", "README.md", "pyproject.toml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
